Fix Bugs

+++ src/public/build/controllers/Headers.controllers.js
@@ -1,11 +1,17 @@
 "use-strict";
 
+import { Accueil } from "../views/Accueil.views.js";
 import { Stats } from "../views/Stats.views.js";
 // import { Categories } from "../views/Categories.js"
 
-export function HeaderController() {
+export function HeaderController(username = null) {
   const categories = document.querySelectorAll(".categories");
   const stats = document.querySelectorAll(".stats");
+
+  logo.addEventListener("click", () => {
+    localStorage.removeItem("Quizz");
+    Accueil(username);
+  });
 
   stats.forEach((stat) => {
     stat.addEventListener("click", () => {
@@ -19,8 +25,10 @@ export function HeaderController() {
     });
   });
 
-  logout.addEventListener("click", () => {
-    localStorage.clear();
-    document.location.reload("http://127.0.0.1:5500/src/public/");
-  });
+  if (username) {
+    logout.addEventListener("click", () => {
+      localStorage.clear();
+      document.location.reload("http://127.0.0.1:5500/src/public/");
+    });
+  }
 }
